pyproject.toml
[project]
name = "discord-jp-sentence-patterns"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.6.4",
    "dotenv>=0.9.9",
    "google-generativeai>=0.8.6",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "supabase>=2.27.0",
]
